rename to types.ts for consistency
