+++ packages/jest-environment/setup.ts
@@ -1,1 +1,1 @@
-import "./";
+import "./src/types";
