+++ packages/jest-environment/src/index.ts
@@ -1,3 +1,4 @@
 import { RunnerEnvironment } from "./RunnerEnvironment";
+import "./types";
 
 module.exports = RunnerEnvironment;
